requirements.txt
langgraph>=0.2.24
numpy>=1.24
pandas>=2.0
PyYAML>=6.0
scipy>=1.10
matplotlib>=3.7
